+++ D3dxSkinManager.Client/src/modules/help/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Help Module
+ * Exports help-related components
+ */
+
+export { HelpWindow } from './components/HelpWindow';